# sechub-api-java/src/main/resources/reduced-openapi3.json
{
    "components": {
        "schemas": {
...
            "ExecutorConfiguration": {
                "title": "ExecutorConfiguration",
                "type": "object",
                "properties": {
                    "productIdentifier": {
                        "type": "string",
                        "description": "Executor product identifier"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of this configuration"
                    },
                    "setup": {
                        "type": "object",
                        "properties": {
                            "baseURL": {
                                "type": "string",
                                "description": "Base URL to the product"
                            },
                            "credentials": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "description": "Password, either plain (not recommended) or with env:VARIABLENAME, in last case the password will be from environment variable "
                                    },
                                    "user": {
                                        "type": "string",
                                        "description": "User name, either plain (not recommended) or with env:VARIABLENAME, in last case the user name will be from environment variable "
                                    }
                                }
                            },
                            "jobParameters": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "value": {
                                            "type": "string",
                                            "description": "Job parameter value"
                                        },
                                        "key": {
                                            "type": "string",
                                            "description": "Job parameter key"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "executorVersion": {
                        "type": "number",
                        "description": "Executor version"
                    },
                    "enabled": {
                        "type": "boolean",
                        "description": "Enabled state of executor, per default false"
                    }
                }
            },
            "ExecutorConfigurationId": {
                "title": "ExecutorConfigurationId",
                "type": "object"
            },
            "ExecutorConfigurationWithUUID": {
                "title": "ExecutorConfigurationWithUUID",
                "type": "object",
                "properties": {
                    "productIdentifier": {
                        "type": "string",
                        "description": "Executor product identifier"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of this configuration"
                    },
                    "setup": {
                        "type": "object",
                        "properties": {
                            "baseURL": {
                                "type": "string",
                                "description": "Base URL to the product"
                            },
                            "credentials": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "description": "Password, either plain (not recommended) or with env:VARIABLENAME, in last case the password will be from environment variable "
                                    },
                                    "user": {
                                        "type": "string",
                                        "description": "User name, either plain (not recommended) or with env:VARIABLENAME, in last case the user name will be from environment variable "
                                    }
                                }
                            },
                            "jobParameters": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "value": {
                                            "type": "string",
                                            "description": "Job parameter value"
                                        },
                                        "key": {
                                            "type": "string",
                                            "description": "Job parameter key"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "executorVersion": {
                        "type": "number",
                        "description": "Executor version"
                    },
                    "uuid": {
                        "type": "string",
                        "description": "The uuid of this configuration"
                    },
                    "enabled": {
                        "type": "boolean",
                        "description": "Enabled state of executor"
                    }
                }
            },
...
            "ListOfExecutorConfigurations": {
                "title": "ListOfExecutorConfigurations",
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "description": "Always `executorConfigurationList` as an identifier for the list"
                    },
                    "executorConfigurations": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "description": "The configuration name"
                                },
                                "uuid": {
                                    "type": "string",
                                    "description": "The uuid of the configuration"
                                },
                                "enabled": {
                                    "type": "boolean",
                                    "description": "Enabled state of configuration"
                                }
                            }
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "info": {
        "title": "SecHub API",
        "description": "SecHub API description",
        "version": "0.0.0"
    },
    "openapi": "3.0.1",
    "paths": {
...
        "/api/admin/config/executor": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Admin creates an executor configuration",
                "description": "An administrator creates an executor a new configuration entry.",
                "operationId": "adminCreatesExecutorConfiguration",
                "requestBody": {
                    "content": {
                        "application/json;charset=UTF-8": {
                            "schema": {
                                "$ref": "#/components/schemas/ExecutorConfiguration"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "201",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExecutorConfigurationId"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic": [
                            
                        ]
                    }
                ]
            }
        },
        "/api/admin/config/executor/{uuid}": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Admin fetches executor configuration",
                "description": "An administrator fetches one explicit executor configuration by its uuid.",
                "operationId": "adminFetchesExecutorConfiguration",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The configuration uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "200",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExecutorConfigurationWithUUID"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic": [
                            
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "admin"
                ],
                "summary": "Admin updates executor configuration setup",
                "description": "An administrator updateds dedicated executor configuration. The update does change description, enabled state and also used executors, but Will NOT change any associations between profile and projects.",
                "operationId": "adminUpdatesExecutorConfig",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The configuration uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=UTF-8": {
                            "schema": {
                                "$ref": "#/components/schemas/ExecutorConfiguration"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "200"
                    }
                },
                "security": [
                    {
                        "basic": [
                            
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "Admin deletes executor configuration",
                "description": "An administrator deletes an executor by removing the configuration entry identified by its uuid",
                "operationId": "adminDeletesExecutorConfiguration",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The configuration uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "200"
                    }
                },
                "security": [
                    {
                        "basic": [
                            
                        ]
                    }
                ]
            }
        },
        "/api/admin/config/executors": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Admin fetches executor configuration list",
                "description": "An administrator fetches executor configuration list which contains all executor configurations",
                "operationId": "adminFetchesExecutorConfigurationList",
                "responses": {
                    "200": {
                        "description": "200",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListOfExecutorConfigurations"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic": [
                            
                        ]
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "https://localhost"
        }
    ],
...
}